Add better description for Vercel OTEL hack
